[client] ディレクトリ構造修正
