+++ client/src/Product/pages/CreateProduct/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import axios from 'axios'
 import { useHistory } from 'react-router-dom'
-import ProductForm from '../ProductForm'
+import ProductForm from '../../components/ProductForm'
 import PageTitle from 'src/shared/PageTitle'
 
 const CreateProduct: React.SFC = () => {
