+++ client/src/Specification/index.tsx
@@ -4,12 +4,12 @@ import {
   Route,
   useRouteMatch,
 } from "react-router-dom"
-import SpecificationItem from './SpecificationItem'
+import SpecificationItem from './SpecificationItem/SpecificationItemList'
 import CreateSpecificationItem from './SpecificationItem/CreateSpecificationItem'
 import EditSpecificationItem from './SpecificationItem/EditSpecificationItem'
-import SpecificationList from './SpecificationList'
-import CreateSpecification from './CreateSpecification'
-import EditSpecification from './EditSpecification'
+import SpecificationList from './pages/SpecificationList'
+import CreateSpecification from './pages/CreateSpecification'
+import EditSpecification from './pages/EditSpecification'
 
 const Specification: React.SFC = () => {
   let match = useRouteMatch()
